[20180730]_[用户权限模块]

+++ src/main/java/com/vip/darker/controller/UserController.java
@@ -4,9 +4,15 @@ import com.baomidou.mybatisplus.mapper.EntityWrapper;
 import com.baomidou.mybatisplus.plugins.Page;
 import com.vip.darker.model.*;
 import com.vip.darker.system.locator.SystemServiceLocator;
+import com.vip.darker.util.BeanToMapUtil;
+import com.vip.darker.util.Constant;
 import org.springframework.web.bind.annotation.*;
 
+import java.beans.IntrospectionException;
+import java.lang.reflect.InvocationTargetException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @Auther: Darker
@@ -26,20 +32,22 @@ public class UserController {
      * @date: 2018/7/19 22:38
      */
     @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public boolean addUser(@RequestParam(value = "roleId") Integer roleId, UserModel userModel) {
-
-        try {
-            // 用户新增
-            SystemServiceLocator.getUserService().insert(userModel);
+    public Map<String, Object> addUser(@RequestParam(value = "roleId", required = false, defaultValue = "1") Integer roleId, UserModel userModel) {
+        // 用户新增
+        boolean flag = SystemServiceLocator.getUserService().insert(userModel);
+        if (flag) {
             // 用户角色关系数据新增
             URRelation relation = new URRelation();
             relation.setUserId(userModel.getId());
             relation.setRoleId(roleId);
-            SystemServiceLocator.getURRelationService().insert(relation);
-        } catch (Exception e) {
-            return false;
+            flag = SystemServiceLocator.getURRelationService().insert(relation);
         }
-        return true;
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "新增成功!" : "新增失败!");
+
+        return map;
+
     }
 
     /**
@@ -51,20 +59,21 @@ public class UserController {
      * @date: 2018/7/19 22:40
      */
     @RequestMapping(value = "/update", method = RequestMethod.PUT)
-    public boolean updateUser(@RequestBody Integer roleId, @RequestBody UserModel userModel) {
-
-        try {
-            // 用户更新
-            SystemServiceLocator.getUserService().updateById(userModel);
+    public Map<String, Object> updateUser(Integer roleId, UserModel userModel) {
+        // 用户更新
+        boolean flag = SystemServiceLocator.getUserService().updateById(userModel);
+        if (flag) {
             // 用户角色关系数据更新
             URRelation relation = new URRelation();
             relation.setUserId(userModel.getId());
             relation.setRoleId(roleId);
-            SystemServiceLocator.getURRelationService().update(relation, new EntityWrapper<URRelation>().where("userId={0}", userModel.getId()));
-        } catch (Exception e) {
-            return false;
+            flag = SystemServiceLocator.getURRelationService().update(relation, new EntityWrapper<URRelation>().where("userId={0}", userModel.getId()));
         }
-        return true;
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "更新成功!" : "更新失败!");
+
+        return map;
     }
 
     /**
@@ -76,17 +85,39 @@ public class UserController {
      * @date: 2018/7/19 22:43
      */
     @RequestMapping(value = "/delete/{id}", method = RequestMethod.DELETE)
-    public boolean deleteUser(@PathVariable(value = "id") Integer id) {
+    public Map<String, Object> deleteUser(@PathVariable(value = "id") Integer id) {
+        // 用户删除
+        boolean flag = SystemServiceLocator.getUserService().deleteById(id);
+        if (flag) {
+            // 用户角色关系数据删除
+            flag = SystemServiceLocator.getURRelationService().delete(new EntityWrapper<URRelation>().where("userId={0}", id));
+        }
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "删除成功!" : "删除失败!");
+
+        return map;
+    }
+
+    /**
+     * 功能描述: 用户实体查询
+     *
+     * @param: [id]
+     * @return: java.util.Map<>
+     * @auther: darker
+     * @date: 2018/7/30 14:17
+     */
+    @RequestMapping(value = "/all/{id}")
+    public Map<String, Object> queryUserById(@PathVariable(value = "id") Integer id) {
+
+        UserModel userModel = SystemServiceLocator.getUserService().selectById(id);
 
         try {
-            // 用户删除
-            SystemServiceLocator.getUserService().deleteById(id);
-            // 用户角色关系数据删除
-            SystemServiceLocator.getURRelationService().delete(new EntityWrapper<URRelation>().where("userId={0}", id));
-        } catch (Exception e) {
-            return false;
+            return BeanToMapUtil.convertBeanToMap(userModel);
+        } catch (IntrospectionException | IllegalAccessException | InvocationTargetException e) {
+            e.printStackTrace();
         }
-        return true;
+        return null;
     }
 
     /**
@@ -98,84 +129,124 @@ public class UserController {
      * @date: 2018/7/19 22:47
      */
     @RequestMapping(value = "/all", method = RequestMethod.GET)
-    public List<UserModel> queryAllUser(
-            @RequestParam(value = "pageNum", required = false, defaultValue = "1")
-                    Integer pageNum,
-            @RequestParam(value = "pageSize", required = false, defaultValue = "10")
-                    Integer pageSize) {
+    public List<UserModel> queryAllUser(@RequestParam(value = "pageNum", required = false, defaultValue = "1") Integer pageNum, @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
         return SystemServiceLocator.getUserService().selectPage(new Page<>(pageNum, pageSize)).getRecords();
+    }
+
+    /**
+     * 功能描述: 用户分页查询,最大页数
+     *
+     * @param: []
+     * @return: java.util.Map<>
+     * @auther: darker
+     * @date: 2018/7/30 10:48
+     */
+    @RequestMapping(value = "/userMaxPage", method = RequestMethod.GET)
+    public Map<String, Object> getUserMaxPage() {
+
+        Map<String, Object> map = new HashMap<>();
+
+        int count = SystemServiceLocator.getUserService().selectCount(new EntityWrapper<>());
+
+        map.put("userMaxPage", (count - 1) / Constant.PAGE_SIZE + 1);
+
+        return map;
     }
 
     /**
      * 功能描述: 角色新增
      *
      * @param: [permissionId, roleModel]
-     * @return: boolean
+     * @return: java.util.Map<>
      * @auther: darker
      * @date: 2018/7/19 22:50
      */
     @RequestMapping(value = "/addRole", method = RequestMethod.POST)
-    public boolean addRole(@RequestParam(value = "permissionId") Integer permissionId, RoleModel roleModel) {
-
-        try {
-            // 角色新增
-            SystemServiceLocator.getRoleService().insert(roleModel);
+    public Map<String, Object> addRole(@RequestParam(value = "permissionId") Integer permissionId, RoleModel roleModel) {
+        // 角色新增
+        boolean flag = SystemServiceLocator.getRoleService().insert(roleModel);
+        if (flag) {
             // 角色权限关系数据新增
             RPRelation relation = new RPRelation();
             relation.setRoleId(roleModel.getId());
             relation.setPermissionId(permissionId);
             SystemServiceLocator.getRPRelationService().insert(relation);
-        } catch (Exception e) {
-            return false;
         }
-        return true;
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "新增成功!" : "新增失败!");
+
+        return map;
     }
 
     /**
      * 功能描述: 角色更新
      *
      * @param: [permissionId, roleModel]
-     * @return: boolean
+     * @return: java.util.Map<>
      * @auther: darker
      * @date: 2018/7/20 11:26
      */
     @RequestMapping(value = "/updateRole", method = RequestMethod.PUT)
-    public boolean updateRole(@RequestBody Integer permissionId, @RequestBody RoleModel roleModel) {
-
-        try {
-            // 角色更新
-            SystemServiceLocator.getRoleService().updateById(roleModel);
+    public Map<String, Object> updateRole(Integer permissionId, RoleModel roleModel) {
+        // 角色更新
+        boolean flag = SystemServiceLocator.getRoleService().updateById(roleModel);
+        if (flag) {
             // 角色权限关系数据更新
             RPRelation relation = new RPRelation();
             relation.setRoleId(roleModel.getId());
             relation.setPermissionId(permissionId);
-            SystemServiceLocator.getRPRelationService().update(relation, new EntityWrapper<RPRelation>().where("roleId={0}", roleModel.getId()));
-        } catch (Exception e) {
-            return false;
+            flag = SystemServiceLocator.getRPRelationService().update(relation, new EntityWrapper<RPRelation>().where("roleId={0}", roleModel.getId()));
         }
-        return true;
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "更新成功!" : "更新失败!");
+
+        return map;
     }
 
     /**
      * 功能描述: 角色删除
      *
      * @param: [id]
-     * @return: boolean
+     * @return: java.util.Map<>
      * @auther: darker
      * @date: 2018/7/20 11:30
      */
     @RequestMapping(value = "/deleteRole/{id}", method = RequestMethod.DELETE)
-    public boolean deleteRole(@PathVariable(value = "id") Integer id) {
+    public Map<String, Object> deleteRole(@PathVariable(value = "id") Integer id) {
+        // 角色删除
+        boolean flag = SystemServiceLocator.getRoleService().deleteById(id);
+        if (flag) {
+            // 角色权限关系数据删除
+            flag = SystemServiceLocator.getRPRelationService().delete(new EntityWrapper<RPRelation>().where("roleId={0}", id));
+        }
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "删除成功!" : "删除失败!");
+
+        return map;
+    }
+
+    /**
+     * 功能描述: 角色实体查询
+     *
+     * @param: [id]
+     * @return: java.util.Map<>
+     * @auther: darker
+     * @date: 2018/7/30 15:31
+     */
+    @RequestMapping(value = "/allRole/{id}")
+    public Map<String, Object> queryRoleById(@PathVariable(value = "id") Integer id) {
+
+        RoleModel roleModel = SystemServiceLocator.getRoleService().selectById(id);
 
         try {
-            // 角色删除
-            SystemServiceLocator.getRoleService().deleteById(id);
-            // 角色权限关系数据删除
-            SystemServiceLocator.getRPRelationService().delete(new EntityWrapper<RPRelation>().where("roleId={0}", id));
-        } catch (Exception e) {
-            return false;
+            return BeanToMapUtil.convertBeanToMap(roleModel);
+        } catch (IntrospectionException | IllegalAccessException | InvocationTargetException e) {
+            e.printStackTrace();
         }
-        return true;
+        return null;
     }
 
     /**
@@ -187,11 +258,7 @@ public class UserController {
      * @date: 2018/7/20 11:34
      */
     @RequestMapping(value = "/allRole", method = RequestMethod.GET)
-    public List<RoleModel> queryAllRole(
-            @RequestParam(value = "pageNum", required = false, defaultValue = "1")
-                    Integer pageNum,
-            @RequestParam(value = "pageSize", required = false, defaultValue = "10")
-                    Integer pageSize) {
+    public List<RoleModel> queryAllRole(@RequestParam(value = "pageNum", required = false, defaultValue = "1") Integer pageNum, @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
         return SystemServiceLocator.getRoleService().selectPage(new Page<>(pageNum, pageSize)).getRecords();
     }
 
@@ -199,39 +266,73 @@ public class UserController {
      * 功能描述: 权限新增
      *
      * @param: [permissionModel]
-     * @return: boolean
+     * @return: java.util.Map<>
      * @auther: darker
      * @date: 2018/7/19 23:04
      */
     @RequestMapping(value = "/addPermission", method = RequestMethod.POST)
-    public boolean addPermission(PermissionModel permissionModel) {
-        return SystemServiceLocator.getPermissionService().insert(permissionModel);
+    public Map<String, Object> addPermission(PermissionModel permissionModel) {
+
+        boolean flag = SystemServiceLocator.getPermissionService().insert(permissionModel);
+
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "新增成功!" : "新增失败!");
+
+        return map;
     }
 
     /**
      * 功能描述: 权限更新
      *
      * @param: [permissionModel]
-     * @return: boolean
+     * @return: java.util.Map<>
      * @auther: darker
      * @date: 2018/7/20 11:37
      */
     @RequestMapping(value = "/updatePermission", method = RequestMethod.PUT)
-    public boolean updatePermission(@RequestBody PermissionModel permissionModel) {
-        return SystemServiceLocator.getPermissionService().updateById(permissionModel);
+    public Map<String, Object> updatePermission(PermissionModel permissionModel) {
+
+        boolean flag = SystemServiceLocator.getPermissionService().updateById(permissionModel);
+
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "更新成功!" : "更新失败!");
+
+        return map;
     }
 
     /**
      * 功能描述: 权限删除
      *
      * @param: [id]
-     * @return: boolean
+     * @return: java.util.Map<>
      * @auther: darker
      * @date: 2018/7/20 11:39
      */
     @RequestMapping(value = "/deletePermission/{id}", method = RequestMethod.DELETE)
-    public boolean deletePermission(@PathVariable(value = "id") Integer id) {
-        return SystemServiceLocator.getPermissionService().deleteById(id);
+    public Map<String, Object> deletePermission(@PathVariable(value = "id") Integer id) {
+
+        boolean flag = SystemServiceLocator.getPermissionService().deleteById(id);
+
+        Map<String, Object> map = new HashMap<>();
+
+        map.put("msg", flag ? "删除成功!" : "删除失败!");
+
+        return map;
+    }
+
+    /**
+     * 功能描述: 权限实体查询
+     *
+     * @param: [id]
+     * @return: com.vip.darker.model.PermissionModel
+     * @auther: darker
+     * @date: 2018/7/30 15:31
+     */
+    @RequestMapping(value = "/allPremission/{id}")
+    public PermissionModel queryPermissionById(@PathVariable(value = "id") Integer id) {
+        return SystemServiceLocator.getPermissionService().selectById(id);
     }
 
     /**
@@ -243,11 +344,27 @@ public class UserController {
      * @date: 2018/7/20 11:42
      */
     @RequestMapping(value = "/allPermission", method = RequestMethod.GET)
-    public List<PermissionModel> queryAllPermission(
-            @RequestParam(value = "pageNum", required = false, defaultValue = "1")
-                    Integer pageNum,
-            @RequestParam(value = "pageSize", required = false, defaultValue = "10")
-                    Integer pageSize) {
+    public List<PermissionModel> queryAllPermission(@RequestParam(value = "pageNum", required = false, defaultValue = "1") Integer pageNum, @RequestParam(value = "pageSize", required = false, defaultValue = "10") Integer pageSize) {
         return SystemServiceLocator.getPermissionService().selectPage(new Page<>(pageNum, pageSize)).getRecords();
+    }
+
+    /**
+     * 功能描述: 权限分页查询,最大页数
+     *
+     * @param: []
+     * @return: java.util.Map<>
+     * @auther: darker
+     * @date: 2018/7/30 10:48
+     */
+    @RequestMapping(value = "/permissionMaxPage", method = RequestMethod.GET)
+    public Map<String, Object> getPermissionMaxPage() {
+
+        Map<String, Object> map = new HashMap<>();
+
+        int count = SystemServiceLocator.getUserService().selectCount(new EntityWrapper<>());
+
+        map.put("permissionMaxPage", (count - 1) / Constant.PAGE_SIZE + 1);
+
+        return map;
     }
 }
